+++ src/Parser/Priest/Holy/Modules/Spells/HolyWords/HolyWordSerenity.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import HolyWordBase from './HolyWordBase';
+
+class HolyWordSerenity extends HolyWordBase {
+  constructor(...args) {
+    super(...args);
+
+  }
+}
+
+export default HolyWordSerenity;
